ui/requirements.txt
streamlit==1.32.0
pandas
numpy
plotly
python-dotenv
requests
anthropic 